requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.339
python-dotenv==1.0.0
pydantic==2.5.2
openai==1.3.5
python-multipart==0.0.6
tqdm==4.66.1
typer==0.9.0
faiss-cpu==1.7.4
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
unstructured==0.10.30
pyPDF2==3.0.1
docx2txt==0.8
pandas==2.1.3
pytest==7.4.3